完成 后台文章管理

+++ src/main/java/com/vtkd/ssm/blog/service/impl/TagServiceImpl.java
@@ -1,8 +1,11 @@
 package com.vtkd.ssm.blog.service.impl;
 
+import com.github.pagehelper.PageInfo;
 import com.vtkd.ssm.blog.entity.Tag;
+import com.vtkd.ssm.blog.mapper.ArticleTagRefMapper;
 import com.vtkd.ssm.blog.mapper.TagMapper;
 import com.vtkd.ssm.blog.service.TagService;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -14,44 +17,109 @@ import java.util.List;
  * @author 君上
  * @date 2022-7-25
  */
+@Slf4j
 @Service
 public class TagServiceImpl implements TagService {
-
     @Autowired
     private TagMapper tagMapper;
 
-    @Override
-    public void insertTag(Tag tag) {
-        tagMapper.insertTag(tag);
-    }
+    @Autowired
+    private ArticleTagRefMapper acMapper;
 
     @Override
     public void deleteTagById(Integer tagId) {
-        tagMapper.deleteTagById(tagId);
+        try {
+            // 先判断 有没有文章,没有文章才能删除
+            Integer articleCount = acMapper.countArticleByTagId(tagId);
+            if (articleCount ==  0){
+                tagMapper.deleteTagById(tagId);
+            }
+        } catch (Exception e) {
+            log.error("标签删除失败，tagId:{}, cause:{}", tagId, e);
+        }
     }
 
     @Override
     public void updateTag(Tag tag) {
-        tagMapper.updateTag(tag);
+        try {
+            tagMapper.updateTag(tag);
+        } catch (Exception e) {
+            log.error("标签修改失败，tag:{}, cause:{}", tag, e);
+        }
+    }
+
+    @Override
+    public void insertTag(Tag tag) {
+        try {
+            tagMapper.insertTag(tag);
+        } catch (Exception e) {
+            log.error("标签插入失败，tag:{}, cause:{}", tag, e);
+        }
     }
 
     @Override
     public Tag getTagById(Integer tagId) {
-        return tagMapper.getTagById(tagId);
+        Tag tag = null;
+        try {
+            tag = tagMapper.getTagById(tagId);
+        } catch (Exception e) {
+            log.error("根据id获取标签失败，tagId:{}, cause:{}", tagId, e);
+        }
+
+        return tag;
     }
 
     @Override
     public Tag getTagByName(String tagName) {
-        return tagMapper.getTagByName(tagName);
+        Tag tag = null;
+        try {
+            tag = tagMapper.getTagByName(tagName);
+        } catch (Exception e) {
+            log.error("根据name获取标签失败，tagName:{}, cause:{}", tagName, e);
+        }
+
+        return tag;
     }
 
     @Override
     public List<Tag> listTag() {
-        return tagMapper.listTag();
+
+        List<Tag> tagList = null;
+        try {
+            tagList = tagMapper.listTag();
+        } catch (Exception e) {
+            log.error("获取所有标签失败, cause:{}", e);
+        }
+        return tagList;
+    }
+
+    @Override
+    public PageInfo<Tag> pageListTag(Integer pageSize, Integer pageIndex) {
+        return null;
+    }
+
+    @Override
+    public List<Tag> listTagWithCount() {
+
+        List<Tag> tagList = null;
+
+        try {
+
+            tagList = this.listTag();
+            for (Tag tag : tagList) {
+                Integer articleCount = acMapper.countArticleByTagId(tag.getTagId());
+                tag.setArticleCount(articleCount);
+            }
+        } catch (Exception e){
+            e.printStackTrace();
+            log.error("获取所有标签和文章数量失败, cause:{}", e);
+        }
+
+        return tagList;
     }
 
     @Override
     public Integer countTag() {
-        return tagMapper.countTag();
+        return null;
     }
 }
